Reuse move action buttons
