+++ src/app/item-actions/ItemMoveLocations.m.scss.d.ts
@@ -1,0 +1,15 @@
+// This file is automatically generated.
+// Please do not change this file!
+interface CssExports {
+  'disabled': string;
+  'equip': string;
+  'move': string;
+  'moveLocationIcons': string;
+  'moveLocationPadding': string;
+  'moveLocations': string;
+  'moveWithVault': string;
+  'vaultButton': string;
+  'vaultLabel': string;
+}
+export const cssExports: CssExports;
+export default cssExports;
