--- src/app/item-popup/ItemActionButton.m.scss
@@ -1,53 +1,0 @@
-@import '../variables.scss';
-
-.button {
-  > div {
-    background-size: cover;
-    width: 44px;
-    height: 44px;
-    text-align: center;
-    line-height: 10px;
-    // Auto-scale on mobile to fit screen width
-    @include phone-portrait {
-      width: calc((100vw - 20px - 9 * 4px) / 7);
-      height: calc((100vw - 20px - 9 * 4px) / 7);
-    }
-  }
-  margin-right: 4px;
-  cursor: pointer;
-  user-select: none;
-  &:last-child {
-    margin-right: 0;
-  }
-
-  &:focus {
-    outline-width: 5px;
-  }
-
-  span {
-    text-transform: uppercase;
-    text-shadow: 1px 1px 3px rgba(0, 0, 0, 1);
-    background: rgba(0, 0, 0, 0.5);
-    width: 100%;
-    height: 100%;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    &:hover {
-      background: rgba(0, 0, 0, 0.1);
-    }
-  }
-}
-
-.locations {
-  display: flex;
-  flex-direction: row;
-}
-
-.locationsV {
-  display: flex;
-  flex-direction: column;
-  & > * {
-    margin: 4px 0;
-  }
-}
