--- src/app/item-popup/ItemActionButton.m.scss.d.ts
@@ -1,9 +1,0 @@
-// This file is automatically generated.
-// Please do not change this file!
-interface CssExports {
-  'button': string;
-  'locations': string;
-  'locationsV': string;
-}
-export const cssExports: CssExports;
-export default cssExports;
