--- src/app/item-popup/ItemActions.m.scss
@@ -1,34 +1,0 @@
-.interaction {
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-}
-
-.infusePerk {
-  border-radius: 50%;
-  background-size: 80%;
-  background-repeat: no-repeat;
-  background-position: center;
-  background-color: #4887ba;
-}
-
-.weapons {
-  background-image: url('../../images/Weapons.png');
-}
-.armor {
-  background-image: url('../../images/Armor.png');
-}
-
-.destiny2 {
-  background-size: 100%;
-  border-radius: 0;
-  background-color: transparent;
-  background-image: url('../../images/d2infuse.png');
-}
-
-.moveDistribute {
-  background-image: url('../../images/arrows-out.png');
-}
-.moveConsolidate {
-  background-image: url('../../images/arrows-in.png');
-}
