--- src/app/item-popup/ItemActions.m.scss.d.ts
@@ -1,13 +1,0 @@
-// This file is automatically generated.
-// Please do not change this file!
-interface CssExports {
-  'armor': string;
-  'destiny2': string;
-  'infusePerk': string;
-  'interaction': string;
-  'moveConsolidate': string;
-  'moveDistribute': string;
-  'weapons': string;
-}
-export const cssExports: CssExports;
-export default cssExports;
