+++ src/app/item-popup/ItemMoveAmount.scss
@@ -1,3 +1,7 @@
+.move-amount-container {
+  flex: 1;
+}
+
 .move-amount {
   display: flex;
   flex-direction: row;
