--- src/app/item-popup/ItemMoveLocation.m.scss
@@ -1,3 +1,0 @@
-.moveVault {
-  background-image: url('../../images/vault.svg');
-}
