--- src/app/item-popup/ItemMoveLocation.m.scss.d.ts
@@ -1,7 +1,0 @@
-// This file is automatically generated.
-// Please do not change this file!
-interface CssExports {
-  'moveVault': string;
-}
-export const cssExports: CssExports;
-export default cssExports;
