+++ src/app/item-popup/ItemPopupContainer.m.scss.d.ts
@@ -10,6 +10,7 @@ interface CssExports {
   'exotic': string;
   'legendary': string;
   'mobileItemActions': string;
+  'mobileMoveLocations': string;
   'movePopupDialog': string;
   'popupBackground': string;
   'rare': string;
